refactor: Change `UnivariatePolynomilal`'s `degree` method to return `usize` rather than `Option<usize>`
For a while I was treating the degree of the zero polynomial as being distinct from the degree of a constant polynomial. This is mathematically true, but does not appear to be relevant for SumCheck, so I changed this interface a few commits ago for multivariate polynomials. I just realized I forgot to change it for univariate polynomials, hence this change.